Turn playaround into some actual software.
We added a very simple feature first, but left the steps as pending. Then we wrote a functional test which drove out a Gmail 'object', and wrote tests for that. Then we refactored that object into a class, and further refactored it to use smaller descriptive methods. Finally we returned to the feature, and implemented its steps using the Gmail class.