Pass the IMAP connection into the GmailImapClient initializer
I felt that the client shouldn't be responsible for both instantiating and using the IMAP connection.  I also thought that as an IMAP connection has to be authenticated before use, that it should be impossible to have a client where login has yet to be called on the connection.

N.B. Making this change has broken the behaviour of the GmailAccount class, but not caused any tests to fail.  I will fix this on the next commit, and am noting it here because I think it asks interesting questions about the value of our current acceptance/integration tests.  Also, I'm a little bit unclear as the the role of GmailAccount.  It doesn't seem to have anything (yet) to do with gmail, just conversion of messages from raw text to Mail objects.  Am sure this is only a transitory stage though, and things will become clearer.