Introduce a method to import messages given credentials.
We need to be able to import messages from an account via a high-level
method that's suitable for calling from (for example) a Rake task.

We've put this method in `MessageImporter`, but wanted to keep the tests
slightly separate because we weren't 100% sure that it belonged there.

Having written the tests, and with the two `require`s that we've had to
add to `message_importer.rb`, we're now confident that it doesn't belong
in that class; we'll extract it into something else next.