Use Mail to bypass encoding issues
Attempting to import all the emails from tom.ward@gofreerange.com I encountered some nasty encoding problems like this one:

  Encoding::UndefinedConversionError: "\xA3" from ASCII-8BIT to UTF-8
    from /Users/tomw/Projects/freerange/sauron/lib/file_based_message_store.rb:14:in `write'
    from /Users/tomw/Projects/freerange/sauron/lib/file_based_message_store.rb:14:in `[]='

Rather than worry too much about this, I decided to pipe all imported messages through Mail.new to see if that fixed the problem.  It seemed to, but we should be aware of encoding problems that may arise and fix the underlying problem when or if it becomes the next most important thing to do.