Disable Rails' asset pipeline.
We want to run the app in the production environment on linode but the
asset pipeline is causing us problems because it's expecting
pre-compiled assets to exist, yet recap isn't configured to generate
anything.

We're not using any CSS at the moment so the simplest thing to do is
disable it.