Explore not using autoload.
I'm interested in what the consequences of avoiding autoload are. I believe this will help avoid having to load the Rails environment in some situations (like Rake tasks).

I also think that it's a good top-level indicator of coupling; every class that references a constant (e.g. another class) will typically have to require the file that defines that constant, giving a quick indication of how tightly coupled the class is - the more requires, the more coupled, maybe?