Use an ActiveRecord model to record each imported message.
To display a list of messages, there's no need to store them as files on disk.  Only their subject, date and from is required, so store these as records in the database.

One advantage of storing the messages on disk is that it avoids having to reimport data (which can be very slow).  However, using the CachedConnection to IMAP should negate this.