Add the ability to show a full message.
In this case, full takes its most extreme meaning - the entire content of the original email.  This is more as proof that we can still access original message content via the web than anything else.

In order to prevent the display of a list of messages taking ages, the original message is only loaded at the point it is needed, not when returning one or a number of messages from the repository.