Do `let`, `subject` (and my newly invented `stub`) help make tests clearer?
On my personal projects I usually use `let` and `subject` in test cases to lazily declare stubs and other values common to all tests.  For me this makes them easier to read, as there's less cruft in each example.

To try this out, I've copied the definitions for `let` and `subject` from minitest/spec, and gone through some of the test cases converting them to use these new features.  As I was doing this I noticed I was writing lines like `let(:connection) { stub('connection') }` a lot, so I added another convenience method `stub` to do this.

On the whole I prefer the new form of these tests.  I think they clearly show which dependencies are common to most of the test cases.  If an individual example doesn't use these dependencies, maybe it's small indicator that it belongs elsewhere?  Also, if the list of stubs declared becomes large, then that surely shows the object has too many dependencies.  If they are split amongst every example, it may not be so clear.