Compare tests written in rspec and minitest/test::unit.
I've taken a single test (for the AuthenticatedConnection) and rewritten it in a more nested RSpecese.  I've not run the rspec one - it's for illustrative purposes only!

I'm a big fan of rspec, but I'd like to understand more why I prefer it to test/unit.  It's certainly not some of the magic features such as turning `it { is_valid }` into an example.  I avoid stuff like that always.  I often avoid the `should be_draft` stuff too, as often I feel it's not clear.

What I do like, however are the tool support (running specs by line-number makes editor support work exceptionally well), the nested output format[1] which I find easy to scan to understand how an object has been tested, and the messages when a test fails, which output the source code around the test failure.  I also like being able to nest tests, hence this experiment.

Often there are tests which are related by class, but not really testing the same thing.  A particularly nasty example are rails controllers, where each action (in my mind) should be tested in its own test group.  Another example in the test I've is the distinction between constructor methods and using the actual class.  To me these are distinct things; how to get hold of an instance of a class, and what you can do with it once you have one.  However, it's nice to have them in the same parent group.

Anyway, enough rambling.  Thoughts?

[1] Nested output:

GoogleMail::Mailbox::AuthenticatedConnection
  .new(email, password)
    starts a secure connection with the GoogleMail imap server
    authenticates the connection with the given email and password
  (in general)
    remembers the email address it was instantiated with
    delegates calls to the imap connection