Test the Imap Cache without Rails.
Since the default argument for the ActiveSupport cache uses `Rails.root`, I've introduced a simple stub that replicates that behaviour. However, I wonder if it would be better to avoid using Rails in this argument completely - perhaps by having a method we control that can be stubbed in the test environment?