Limit the number of messages from the repository.
We want the web interface to render in a reasonable amount of time. We
could have used something like pagination but given that we don't know
whether we'll ever need to display a list of *all* messages we've
decided to go with this approach for now.

Plus, Tom says, "Pagination is shit. It's the last resort of a
scoundrel."