Ensure the test passes irrespective of source file encoding.
This test was passing for me when run from the terminal but not when run
from TextMate. This is because the string literal "\xA3" (*NOTE* it
doesn't actually matter what that string contains) is encoded as
ASCII-8BIT when the test is run from the terminal, and as UTF-8 when the
test is run from TextMate. The UTF-8 encoding becomes a problem when we
compare the original string to the base64 decoded one that we get back
from our message store. Although the strings contain the same bytes, the
base64 decoded string is encoded as ASCII-8BIT. There's some discussion
about this in this thread[1].

Although by default the test was passing for me when run from the
terminal, it was possible to get it to fail by passing -KU to ruby, e.g.

    ruby -KU -Itest test/unit/file_based_message_store_test.rb

You could also achieve the same failure by manually setting # encoding:
utf-8 at the top of file_based_message_store_test.rb.

[1] http://lists.zenspider.com/pipermail/ruby/2011-August/005785.html