Remove some knowledge of ActiveRecord from the MessageRepository.
The MessageRepository shouldn't have to know how to use ActiveRecord in
order to check for the existence of a given message. The Record object
should nicely encapsulate that for us in the .message_exists? method.

After some discussion, we decided to add two tests (one for the presence
and one for the absence of messages) around Record.message_exists? We've
stuck with a single test around MessageRepository#exists?

This change was inspired by @floehopper's pull request 16 -
https://github.com/freerange/sauron/pull/16.