Ignore incorrectly encoded Next Line unicode characters.
We're replacing the invalid Next Line control character[1] with an empty
string because we believe it is not a printable character.

The order of our gsubs in MailWrapper#subject are important because when
the pound replacement happens, the UTF-8 String containing the '£'
character is no longer compatible with ASCII-8BIT strings, and a
subsequent `gsub` using an ASCII-8BIT string will fail. Because the Next
Line `gsub` doesn't insert any characters, the resulting string is still
compatible with ASCII-8BIT.

[1]: http://www.fileformat.info/info/unicode/char/85/index.htm