Make our walking skeleton test more robust.
The previous implementation of the 'they should be visible on the
messages page' step wasn't doing what we'd intended it to do: It was
asserting that there were elements with classes of .subject, .date and
.sender within an element with a class of .message. And, because
`within` only finds the first matching element it was asserting the same
thing twice.

The fix was to specify which .message container we expected each set of
attributes to appear in, and to specify the text value that we should
see on the page.

Because messages are now displayed in reverse chronological order I've
changed the text of the subjects in our test to make it clearer that the
newer message appears before the older message.

I also temporarily reverted 3d8e286ca76170a248291e800985ef607f7b1559 in
order to check that this more robust implementation would've caught the
problem @lazyatom and I fixed in that commit. It did indeed catch it :-)