Some message bodies apparently can never be downloaded.
When trying to import messages from james.mead@gofreerange.com, it failed on UID 963. Via the console I could load the headers for this message, but no IMAP command would return the message body.

Rather than just leave this message out, I thought it better to include the data we *can* get (headers), and indicate in the body that the message wasn't available. This should at least give some pointers if people are looking for something that the message referred to.