Exclude duplicate messages from the web interface.
We're currently showing each message up to four times (once for each of
our accounts) in the message index. We're excluding duplicates by
grouping by 'subject', 'date' and 'message_id'. The message_id of all
the current messages in the system will be NULL. We'll add a separate
script to populate the historic messages later.