Returning a Message class from a Mailbox removes the need to repeat arguments
When importing a message, multiple methods take the same three arguments: the account, the uid and the raw message content.  By introducing a new class that includes all three of these, these interfaces become much simpler.  This class can also take responsibility for dealing with strange subject encodings - it's likely that the Mailbox knows the most about any encoding weirdnesses that might exist in the messages it holds, so this feels like a better place.