Record was never a good name for models in our database index.
It's more like an index, so can be renamed ActiveRecordMessageIndex.  The interface to the index is limited to methods such as #message_exists? and #highest_uid, and in fact needn't be an ActiveRecord subclass at all.  It would be easy to swap it out with a different index implementation (providing all the data is reimported!).