Clarify `Mail` vs `Message`.
I want to start dealing with the multiple copies of an email that we
might end up importing in a sensible way. I believe that we can refer to
a `Mail` as being a received version of a `Message`. While there are
multiple `Mails`, they all represent the same `Message`.

I've tried to rename everything to better clarify this. We are now
importing mail, rather than importing messages. The controller, on the
other hand, is displaying messages, rather than mails.

There is still a bit more work to do, particularly splitting out the
`MessageRepository` into a `MailRepository` (which is concerned with
storing and indexing the mails as they come in), and a new
`MessageRepository` (which is concerned with emitting the
slightly-more-abstract `Message` instances which represent one or more
`Mails`). I didn't want to make that change in this commit, as it felt
useful to clarify the terminology change independently from an
architectural change.