Use an identifier to load a Message which isn't the UID of a single Mail
We considered a few approaches, including extracting a Message record
for storage in the database, but decided that the simplest incremental
change was to use a simple identifier that was still stored against the
index record for each Mail.

Some Mails don't have Message IDs (mostly or possibly all from Pivotal
Tracker, where the mail headers contain a blank line that confuses the
Mail gem), so where that does not have a value, we use the sender, date
and subject to generate a stable identifier for the Message.