Introduce an ElasticSearch index to replace the existing ActiveRecord one.
Using ElasticSearch over ActiveRecord has a number of advantages, particularly when it comes to doing full text searches over all messages (not yet implemented).

To replace the existing index, a new class has been added with tests for all of the original interface.  This class can be used by the repository as a drop in replacement.

N.B. Datetime formats were removed from cucumber tests as they were failing due to time format issues, and didn't seem to capture the essence of the features they described.