Fix a problem with importing malformed emails.
We saw the following problem when importing emails for @floehopper's
account:

  undefined method `join' for "tracker-noreply@pivotaltracker.com":String

This error occurred for the email with UID 42111.  Since the Mail gem
doesn't parse this email correctly we don't get a Message ID, which
results in us trying to generate a hash based on a combination of From,
Date and Subject.  We were incorrectly assuming that MailWrapper#from
was returning an array while it actually returns a string or nil.  We've
updated MessageRepository#add_mail to call #to_s instead of #join, and
updated the (somewhat brittle?) MessageRepository tests.