Drastically reduce the number of recent messages displayed.
Because we now call Message#received_by? for every message on the
messages index page, we end up parsing the raw message for
each mail in order to obtain the recipients. This is clearly very inefficient
and the messages index page was taking ages to load / timing out. This
drastic reduction in the number of messages means the page loads in a
reasonable time and draws attention to the underlying performance issue.