A "primary" message index record has many mail index records.
Eventually we want to be able to obtain all the recipients for a given
message by querying its "primary" message index record which in turn
should query all its constituent mail index records. This will be
instead of having to query all of the message index records for the
given message.

This commit makes ActiveRecordMessageIndex#recipients available, but
doesn't use them anywhere.

I felt that I wanted a test that hit the database to check that I had
declared the `has_many` in ActiveRecordMessageIndex correctly.