Use message index record #recipients.
- By changing the setup to pretend that #delivered_to is no longer
available on message index records and to make the data available via
implementation of Message#recipients to obtain the data from the mail
index records on the "primary" message index record.
- The data is actually still available by querying each of the
relevant message index records, we're just pretending it's not for now
so that this commit is a smaller step.
- A Message now only needs access to the "primary" message index record
and not all of its message index records. However we are still
supplying all of them to the Message constructor for now.
- However now we need to ensure that the first messsage index record
passed into the message constructor is the "primary" one. For now this
is achieved by changing the implementation of
ActiveRecordMessageIndex.find_all_by_message_hash to order the message
index records in the same order as
ActiveRecordMessageIndex.find_primary_message_index_record. This wasn't
important before, because we were using all the message index records to
obtain the list of delivered_to addresses and any of the message index
records was sufficient to retrieve a copy of the raw message from the
mail store. But now only the primary message index record will have the
child mail index records which have all the delivered_to addresses.