Return all message IDs that a message is related to.
It can be difficult (and potentially impossible) to reach every message in a
conversation by following the 'In-Reply-To' header alone; in particular, if a
message was not replied to, but a subsequent message was, then there's no
route back to the first reply from the latest message.

It looks like GMail at least is helpful in that it also tracks the message IDs
that are related, and I believe that we can use this to build a more complete
graph of every message in a conversation.