Move responsibility for message hash generation into the index.
The hash can be seen as an identifier of the message as a whole (as opposed to a single email).  As it's the index that combines mails into messages, it makes sense to generate the hash inside the index.

P.S. This is what the elastic search version does, so will make merging easier.