Flag messages that we haven't responded to.
The idea behind this feature is that it would be useful to see at a glance which conversational threads are in progress and could be progressed with a response from us.

Having now implemented this simple approach, it's clearer now that simply checking for replies isn't enough; much of our email traffic comes from services like Pivotal or Travis where a response is not necessarily required. This makes me think that some kind of curation will also be required (perhaps as simple as an overriding flag that says "no, this doesn't need a response", or a whitelist of senders like Pivotal, that are not considered as requiring a response).

I've also had to drastically reduce the number of messages loaded (even 50 makes the page take 8 seconds to load on my laptop). This is as a result of loading messages and then generating conversations from them; it makes me wonder more about implementing a persistent conversation model which is then loaded directly in the controller.