Do we even need to prepare the database?
I think the default rake task already calls db:test:prepare; all we need to do is get the 'development' database ready to have its schema cloned.

The build is still failing, although I'm not sure why; it wants to run the migrations even though the schema is up to date. Perhaps this will fix it, but who knows.