Start moving responsibility for mail parsing to a dedicated class.
For the most part we rely on the mail gem to parse our raw emails.  However, it has a number of oddities and quirks that we need to accommodate.  Previously the code handling these quirks was split between the GoogleMail::Mailbox::Mail class and the MessageRepository::Message.  This will cause us issues as raw emails straight from the mailbox will return different values for various attributes than those read from the mail store.  To rectify this, we have moved all responsibility out of the GoogleMail::Mailbox::Mail class, and will shortly do the same to the MessageRepository::Message class.