Change test to better match the test name.
I believe this test is really concerned with hinting to the mailbox that the importer doesn't want to deal with messages it has already seen. It does this by querying the repository, and then passing that value to the mailbox.

If the above is correct, then the real behaviour to verify is whether or not that value is passed to the the call to `mailbox.uids`.