Add another Pivotal Tracker email address to the exclusions.
This probably won't be useful for new messages, but might be of use for
older ones. I'm not planning on re-importing all the messages just for
this change, because I don't think it's worthwhile. However, I suspect
they will get re-imported again soon so that will take care of things.