Add the conversation date to the list of conversations.
We've chosen to implement the microformat datetime design pattern[1]
which makes it easier to write a stable test without constraining the
human readable date format.

We've introduced the conversation_stub helper method in the
ConversationsControllerTest. We plan to rationalise the different
mail|message|conversation stub factories in a later commit.

[1] http://microformats.org/wiki/datetime-design-pattern