Giving stubs explicit names makes tests easier to debug.
This is because if a stub receives an unexpected method call, or an
expectation on a stub isn't satisfied, it is easier to determine exactly
which stub object is complaining.

If we make the name optional, it's unlikely that we'll give stubs names,
and so won't see any of this benefit; that's why the parameter is
currently mandatory. We have an idea for a compromise, which we might
explore in a later commit.