Halve time taken for `ElasticSearchMessageIndexTest` to run.
This change reduces the time taken from ~30 seconds to ~15 seconds.

It's also probably better that the number of messages to show is a
concern of the controller, rather than a fixed setting of the index.

(This change also makes it possible to iterate over *every* message, by
providing a very very very large limit. This might be useful for
re-processing messages in the future.)