Show the participants of a conversation.
This lets us quickly see who was involved in a conversation.

It's likely that this presentation will quickly be problematic, but I'd
prefer to feel the pain of many participants before implementing a fix,
even if that doesn't take very long at all.

I've taken this opportunity to refactor the interface to
`Conversation.new` to simply take a record; the record should respond to
all of the information that we require for the index page (`subject`,
`date`, `participants`) for fast access.