First steps towards showing a conversation.
In order to do this, we need to be able to load individual conversations, and link to them in the list view.

To take advantage of Rails linking helpers, the `ConversationRepository` needs to return objects that implement `ActiveModel::Naming`. This has driven out the return of actual `Conversation` instances, rather than passing through the `ConversationRecord` that is actually returned by the `ActiveRecordStore`. The  functional tests around `ConversationController` are accordingly modified to return `Conversation` instances wrapped around stub records.