Fix build on Travis due to differing Ruby behaviour.
For whatever reason, when Travis switched to Ruby 1.9.3-p194 on their working
machines, the test in `test/unit/conversation_repository_test.rb` failed[1]:

    1) Error:
    test_retrieves_the_most_recent_conversations_from_the_conversation_index(ConversationRepositoryTest):
    ArgumentError: wrong number of arguments(1 for 0)
        /home/vagrant/builds/freerange/sauron/test/unit/conversation_repository_test.rb:8:in `initialize'
        /home/vagrant/builds/freerange/sauron/test/unit/conversation_repository_test.rb:8:in `new'
        /home/vagrant/builds/freerange/sauron/test/unit/conversation_repository_test.rb:8:in `block in <class:ConversationRepositoryTest>'
        /home/vagrant/.rvm/gems/ruby-1.9.3-p194/gems/mocha-0.10.5/lib/mocha/integration/mini_test/version_230_to_262.rb:28:in `run'

The only change that I could detect was the bump in Ruby patch-levels.

I could not replicate this error locally, so I installed Ruby 1.9.3-p194 in an
Ubuntu 11.10 32-bit VM (which is what Travis uses), and thankfully could
replicate the issue.

It looks like the test in
`test/unit/conversation_repository/conversation_test.rb` was being loaded, and
the fact that the test was put in the `ConversationRepository` namespace by
re-opening the class meant that the real `ConversationRepository` constant in
`lib/conversation_repository.rb` was never being loaded by Rails, because the
constant had been defined.

It's not clear why this was the only test that failed, because this test style
(re-opening the application class) is used in most of the other tests.

In the past we have discussed the potential issues with putting tests inside
application classes, and relying purely on Rails' autoloading rather than
requiring files; perhaps this is a concrete case in support of revising those
positions?

[1]: http://travis-ci.org/#!/freerange/sauron/builds/1676130